fix: move app count tracking to a separate module
