+++ src/components/Dashboard/SafeApps/__tests__/utils.test.ts
@@ -1,5 +1,5 @@
 import { rankTrackedSafeApps } from 'src/components/Dashboard/SafeApps/utils'
-import { AppTrackData } from 'src/routes/safe/components/Apps/types'
+import { AppTrackData } from 'src/routes/safe/components/Apps/trackAppUsageCount'
 
 describe('rankTrackedSafeApps', () => {
   it('ranks more recent apps higher', () => {
